+++ CHANGELOG.md
@@ -40,6 +40,7 @@ _This release is scheduled to be released on 2024-01-01._
 - Fix #3256 filter out bad results from rrule.between
 - Fix calendar events sometimes not respecting deleted events (#3250)
 - Fix electron loadurl locally on Windows when address "0.0.0.0" (#2550)
+- Fix updatanotification (update_helper.js): catch error if reponse is not an JSON format (check PM2)
 
 ## [2.25.0] - 2023-10-01